# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "sneat",
        "description": "Description de votre API",
        "contact": {
            "name": "Nom du contact",
            "email": "contact@api.com"
        },
        "license": {
            "name": "Licence de l'API",
            "url": "URL de la licence"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/publiez/publiez": {
            "post": {
                "tags": [
                    "Publications"
                ],
                "summary": "Publier un statut dans un groupe",
                "description": "Publiez dans le groupe pour que ca se voit aussi sur actu des amis ou non amis membre",
                "operationId": "publiezInGroup",
                "requestBody": {
                    "description": "Données du statut à publier",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_group": {
                                        "type": "integer",
                                        "example": "bonjour tout le monde"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "bonjour tout le monde"
                                    },
                                    "file": {
                                        "type": "string",
                                        "example": "base 64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut publié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/publier/create": {
            "post": {
                "tags": [
                    "Publications"
                ],
                "summary": "Publier un statut",
                "description": "Publier un statut en envoyant les données du statut dans le corps de la requête.",
                "operationId": "publierStatut",
                "requestBody": {
                    "description": "Données du statut à publier",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "bonjour tout le monde"
                                    },
                                    "file": {
                                        "type": "string",
                                        "example": "dqsfjkqsdfkljqhsfdkjhqsfkj"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut publié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/publier/delete/{id_publication}": {
            "post": {
                "tags": [
                    "Publications"
                ],
                "summary": "Suprimer une publication",
                "description": "Suprimer une publication en envoyant le id_publication",
                "operationId": "supprimerStatut",
                "parameters": [
                    {
                        "name": "id_publication",
                        "in": "path",
                        "description": "ID du publication",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supression de publiction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/share/publication/myMur/{publication}": {
            "post": {
                "tags": [
                    "Publications"
                ],
                "summary": "partager et publier un statut",
                "description": "Publier un statut en partagant une statut d-un autre utilisateur.",
                "operationId": "partageInMyMur",
                "parameters": [
                    {
                        "name": "publication",
                        "in": "path",
                        "description": "ID du publication",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données du statut à publier",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Un status de John Doe que je vient de partager"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut publié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/otherProfil/{user}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Voir une profile d'un autre user",
                "description": "Voir une profile d'un autre user",
                "operationId": "otherProfil",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID du user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voir une profile d'un autre user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer votre_token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": [
                "read",
                "write"
            ]
        }
    ]
}